ServerContract refactor
