+++ common/src/main/kotlin/dev/biserman/wingscontracts/command/LoadContractCommand.kt
@@ -5,8 +5,6 @@ import com.mojang.brigadier.arguments.StringArgumentType
 import com.mojang.brigadier.builder.ArgumentBuilder
 import dev.biserman.wingscontracts.WingsContractsMod
 import dev.biserman.wingscontracts.command.ModCommand.giveContract
-import dev.biserman.wingscontracts.core.AbyssalContract
-import dev.biserman.wingscontracts.core.BoundContract
 import dev.biserman.wingscontracts.core.Contract
 import dev.biserman.wingscontracts.data.ContractDataReloadListener
 import dev.biserman.wingscontracts.data.ContractSavedData
@@ -17,11 +15,6 @@ import net.minecraft.commands.arguments.EntityArgument
 import net.minecraft.world.level.Level
 
 object LoadContractCommand {
-    val options = mapOf<String, (Level, ContractTag) -> Contract>(
-        "abyssal" to { level, tag -> AbyssalContract.load(tag, ContractSavedData.get(level)) },
-        "bound" to { level, tag -> BoundContract.load(tag) }
-    )
-
     fun register(): ArgumentBuilder<CommandSourceStack, *> =
         Commands.literal("load")
             .then(
@@ -66,7 +59,7 @@ object LoadContractCommand {
 
             val tag = ContractTag.fromJson(JsonParser.parseString(jsonString).asJsonObject)
 
-            return options[type]!!.invoke(level, tag)
+            return ContractTypeArgument.options[type]!!.load(tag, ContractSavedData.get(level))
         } catch (e: Error) {
             WingsContractsMod.LOGGER.error(e)
             throw e
